requirements.txt
# Deep Secure - Deepfake Detection Model
# Requirements for training and inference

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars & Utilities
tqdm>=4.65.0

# Face Detection (for preprocessing)
mtcnn>=0.1.1
facenet-pytorch>=2.5.2

# Model Evaluation
scikit-learn>=1.2.0
scipy>=1.10.0

# Jupyter Notebook Support (for Kaggle)
jupyter>=1.0.0
ipython>=8.12.0
ipywidgets>=8.0.0

# Optional: For advanced features
# tensorboard>=2.12.0  # For training visualization
# albumentations>=1.3.0  # For advanced augmentation
# timm>=0.9.0  # For additional model architectures
